SampleCreator.java moved from */common to */common/base and build.xml and other files are alligned accordingly
