+++ templates/server/build.xml
@@ -105,6 +105,7 @@
         <exclude name="*management/*"/>
         <exclude name="*management/**/*"/>
         <exclude name="*common/builders/*"/>
+        <exclude name="*common/test/*"/>
         <!-- exclude application specific Money class and related classes -->
         <exclude name="**/*Money*.java"/>
       </fileset>
